+++ src/cpu_lib/managers/getters/__init__.py
@@ -1,2 +1,3 @@
 """Exposed getters classes and methods."""
 from cpu_lib.managers.getters.temperature_getter import TemperatureGetter
+from cpu_lib.managers.getters.use_getter import UseGetter