+++ tests/mock_cpu_lib/__init__.py
@@ -1,4 +1,4 @@
 """Exposed mock_ram_lib classes and methods."""
-from tests.mock_cpu_lib.mock_cpu_manager import mock_temperature
+from tests.mock_cpu_lib.mock_cpu_manager import mock_temperature, mock_use
 
 mock_command_executor_method = "shell_executor_lib.CommandManager.execute_command"