+++ tests/mock_cpu_lib/mock_cpu_manager.py
@@ -1,3 +1,5 @@
 """Mocks of CPU."""
 
 mock_temperature = ["27800", "29000", "34900"]
+
+mock_use = ["25,4"]